refactor: 안쓰는 클래스, 메서드 제거 / 패키지 분리
