+++ src/main/java/domain/hand/Hand.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.hand;
+
+import domain.card.Card;
 
 import java.util.List;
