+++ src/main/java/domain/hand/ScoreCalculator.java
@@ -1,4 +1,7 @@
-package domain;
+package domain.hand;
+
+import domain.card.Card;
+import domain.card.CardNumber;
 
 import java.util.List;
