+++ src/main/java/domain/participant/GameParticipants.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 public class GameParticipants {
     private final Dealer dealer;
