+++ src/main/java/domain/participant/Player.java
@@ -1,10 +1,12 @@
-package domain;
+package domain.participant;
 
+
+import domain.hand.Hand;
 
 public class Player extends Participant {
     private final Name name;
     private final BettingMoney bettingMoney;
-    private Player(Name name, Hand hand ,BettingMoney bettingMoney) {
+    private Player(Name name, Hand hand , BettingMoney bettingMoney) {
         super(hand);
         this.name = name;
         this.bettingMoney = bettingMoney;
