+++ src/main/java/domain/participant/PlayerCreationInfo.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 public class PlayerCreationInfo {
     private final Name name;
