+++ src/test/java/domain/BettingMoneyTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.participant.BettingMoney;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
