+++ src/test/java/domain/DealerTest.java
@@ -4,7 +4,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
-import org.junit.jupiter.api.DisplayName;
+
+import domain.card.Card;
+import domain.card.CardNumber;
+import domain.card.CardShape;
+import domain.hand.Hand;
+import domain.participant.Dealer;
 import org.junit.jupiter.api.Test;
 
 class DealerTest {
