+++ src/test/java/domain/DeckTest.java
@@ -1,6 +1,6 @@
 package domain;
 
-import org.junit.jupiter.api.DisplayName;
+import domain.card.Deck;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
